+++ include/hl_monitoring/opencv_image_provider.h
@@ -31,7 +31,7 @@ public:
   
   void restartStream() override;
 
-  CalibratedImage getCalibratedImage(double time_stamp) override;
+  CalibratedImage getCalibratedImage(uint64_t time_stamp) override;
 
   void update() override;
   
@@ -41,7 +41,7 @@ public:
 
   void saveVideoMetaInformation();
 
-  double getStart() const override;
+  uint64_t getStart() const override;
 
 private:
   /**
@@ -72,7 +72,7 @@ private:
   /**
    * Provide access to index using time_stamps
    */
-  std::map<double, int> indices_by_time_stamp;
+  std::map<uint64_t, int> indices_by_time_stamp;
 
   /**
    * Index of the next image read in the video
